Change Type of price

+++ Entity/Position.cs
@@ -10,6 +10,6 @@ namespace db_projektarbeit
         public int Id { get; set; }             // Id für Index
         public Product Product  { get; set; }   // Product
         public int Counter { get; set; }        // Anzahl der Itemes
-        public double Total { get; set; }       // gesammt Summe aller Artikel
+        public decimal Total { get; set; }       // gesammt Summe aller Artikel
     }
 }
